List considerations for each message.nack() call
With Cloud Pub/Sub's default retry logic, nacking messages can make the
workers spin on nacks instead of making progress on good messages. For
now, this commit just tries to enumerate the various cases that could
lead to a nack, and whether each case is actually useful to nack or not.
We'll probably follow up with more changes to mitigate the issue.
(Possibly just a README change recommending changing Pub/Sub's retry
logic.)